+++ app/src/main/java/com/vrviu/streamer/SceneDetect.java
@@ -9,6 +9,6 @@ public class SceneDetect {
      }
 
      public native boolean init(String fileName, float threshold, int roiX, int roiY, int roiW, int roiH);
-     public native boolean detect(byte[] pixel, int width, int height);
+     public native boolean detect(int[] pixel, int width, int height);
      public native boolean release();
 }
